Rename tests folder to other_files
Change name of the tests folder to other_files to make the name more accurate and clear.
